Trivial issue: rename files according to class names.
