+++ src/adapters/matrix_armadillo.h
