+++ src/adapters/matrix_eigen.h
