+++ src/adapters/matrix_opencv.h
